Lecture_06/requirements.txt
# Lecture 06: RAG Architecture
# ============================

# Core LangChain
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# OpenAI
openai>=1.50.0

# Vector Stores
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Environment
python-dotenv>=1.0.0

# Text Processing
tiktoken>=0.5.0

# Optional: для Jupyter notebooks
ipykernel>=6.0.0
jupyter>=1.0.0

